backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# AI/Agents
openai>=1.12.0
openai-agents>=0.1.0

# Event streaming
aiokafka>=0.10.0

# HTTP client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
